chore(test): move testing utilities to test_util crate (#6360)
