+++ Cargo.toml
@@ -4,7 +4,8 @@ members = [
   "core",
   "tools/hyper_hello",
   "deno_typescript",
-  "test_plugin"
+  "test_plugin",
+  "test_util",
 ]
 exclude = [
   "std/hash/_wasm"
