+++ cli/http_util.rs
@@ -248,7 +248,7 @@ mod tests {
 
   #[tokio::test]
   async fn test_fetch_string() {
-    let http_server_guard = crate::test_util::http_server();
+    let http_server_guard = test_util::http_server();
     // Relies on external http server. See tools/http_server.py
     let url =
       Url::parse("http://127.0.0.1:4545/cli/tests/fixture.json").unwrap();
@@ -267,7 +267,7 @@ mod tests {
 
   #[tokio::test]
   async fn test_fetch_gzip() {
-    let http_server_guard = crate::test_util::http_server();
+    let http_server_guard = test_util::http_server();
     // Relies on external http server. See tools/http_server.py
     let url = Url::parse(
       "http://127.0.0.1:4545/cli/tests/053_import_compression/gziped",
@@ -291,7 +291,7 @@ mod tests {
 
   #[tokio::test]
   async fn test_fetch_with_etag() {
-    let http_server_guard = crate::test_util::http_server();
+    let http_server_guard = test_util::http_server();
     let url = Url::parse("http://127.0.0.1:4545/etag_script.ts").unwrap();
     let client = create_http_client(None).unwrap();
     let result = fetch_once(client.clone(), &url, None).await;
@@ -316,7 +316,7 @@ mod tests {
 
   #[tokio::test]
   async fn test_fetch_brotli() {
-    let http_server_guard = crate::test_util::http_server();
+    let http_server_guard = test_util::http_server();
     // Relies on external http server. See tools/http_server.py
     let url = Url::parse(
       "http://127.0.0.1:4545/cli/tests/053_import_compression/brotli",
@@ -341,7 +341,7 @@ mod tests {
 
   #[tokio::test]
   async fn test_fetch_once_with_redirect() {
-    let http_server_guard = crate::test_util::http_server();
+    let http_server_guard = test_util::http_server();
     // Relies on external http server. See tools/http_server.py
     let url =
       Url::parse("http://127.0.0.1:4546/cli/tests/fixture.json").unwrap();
@@ -398,13 +398,13 @@ mod tests {
 
   #[tokio::test]
   async fn test_fetch_with_cafile_string() {
-    let http_server_guard = crate::test_util::http_server();
+    let http_server_guard = test_util::http_server();
     // Relies on external http server. See tools/http_server.py
     let url =
       Url::parse("https://localhost:5545/cli/tests/fixture.json").unwrap();
 
     let client = create_http_client(Some(String::from(
-      crate::test_util::root_path()
+      test_util::root_path()
         .join("std/http/testdata/tls/RootCA.pem")
         .to_str()
         .unwrap(),
@@ -424,14 +424,14 @@ mod tests {
 
   #[tokio::test]
   async fn test_fetch_with_cafile_gzip() {
-    let http_server_guard = crate::test_util::http_server();
+    let http_server_guard = test_util::http_server();
     // Relies on external http server. See tools/http_server.py
     let url = Url::parse(
       "https://localhost:5545/cli/tests/053_import_compression/gziped",
     )
     .unwrap();
     let client = create_http_client(Some(String::from(
-      crate::test_util::root_path()
+      test_util::root_path()
         .join("std/http/testdata/tls/RootCA.pem")
         .to_str()
         .unwrap(),
@@ -454,10 +454,10 @@ mod tests {
 
   #[tokio::test]
   async fn test_fetch_with_cafile_with_etag() {
-    let http_server_guard = crate::test_util::http_server();
+    let http_server_guard = test_util::http_server();
     let url = Url::parse("https://localhost:5545/etag_script.ts").unwrap();
     let client = create_http_client(Some(String::from(
-      crate::test_util::root_path()
+      test_util::root_path()
         .join("std/http/testdata/tls/RootCA.pem")
         .to_str()
         .unwrap(),
@@ -486,14 +486,14 @@ mod tests {
 
   #[tokio::test]
   async fn test_fetch_with_cafile_brotli() {
-    let http_server_guard = crate::test_util::http_server();
+    let http_server_guard = test_util::http_server();
     // Relies on external http server. See tools/http_server.py
     let url = Url::parse(
       "https://localhost:5545/cli/tests/053_import_compression/brotli",
     )
     .unwrap();
     let client = create_http_client(Some(String::from(
-      crate::test_util::root_path()
+      test_util::root_path()
         .join("std/http/testdata/tls/RootCA.pem")
         .to_str()
         .unwrap(),
