+++ test_plugin/Cargo.toml
@@ -11,3 +11,6 @@ crate-type = ["cdylib"]
 [dependencies]
 futures = "0.3.5"
 deno_core = { path = "../core" }
+
+[dev-dependencies]
+test_util = { path = "../test_util" }
