+++ test_util/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "test_util"
+version = "0.1.0"
+authors = ["the Deno authors"]
+edition = "2018"
+publish = false
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+lazy_static = "1.4.0"
+os_pipe = "0.9.2"
+regex = "1.3.9"
+tempfile = "3.1.0"
